ViewModel and assets added
