+++ app/build.gradle
@@ -53,6 +53,7 @@ dependencies {
     implementation 'androidx.lifecycle:lifecycle-viewmodel-compose:2.4.1'
     implementation 'androidx.compose.runtime:runtime-livedata:1.1.1'
     implementation 'androidx.compose.runtime:runtime-rxjava2:1.1.1'
+    implementation 'com.google.android.exoplayer:exoplayer:2.18.1'
 
     testImplementation 'junit:junit:4.13.2'
