+++ app/src/main/java/com/silverorange/videoplayer/MainActivity.kt
@@ -3,40 +3,19 @@ package com.silverorange.videoplayer
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material.Text
-import androidx.compose.material.TopAppBar
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
+import androidx.lifecycle.LifecycleOwner
+import androidx.lifecycle.ViewModelProvider
 
-class MainActivity : ComponentActivity() {
-  override fun onCreate(savedInstanceState: Bundle?) {
-    super.onCreate(savedInstanceState)
-    setContent {
-      TopBar()
-      TextCard()
+class MainActivity : ComponentActivity(), LifecycleOwner {
+    private val TAG = "MainActivity"
+    lateinit var mainViewModel: MainViewModel
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        mainViewModel = ViewModelProvider(this)[MainViewModel::class.java]
+        lifecycle.addObserver(mainViewModel)
+        setContent {
+            TopBar()
+            TextCard(mainViewModel)
+        }
     }
-  }
-}
-
-@Preview
-@Composable
-fun TopBar() {
-  TopAppBar(title = { Text(text = "Video Player") })
-}
-
-@Preview
-@Composable
-fun TextCard() {
-  Column(
-    modifier = Modifier.fillMaxSize(),
-    verticalArrangement = Arrangement.Center,
-    horizontalAlignment = Alignment.CenterHorizontally
-  ) {
-    Text(text = "Hello world!")
-  }
 }
