+++ app/src/main/java/com/silverorange/videoplayer/MainView.kt
@@ -1,0 +1,26 @@
+package com.silverorange.videoplayer
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material.Text
+import androidx.compose.material.TopAppBar
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+
+@Composable
+fun TopBar() {
+    TopAppBar(title = { Text(text = "Video Player") })
+}
+
+@Composable
+fun TextCard(mainViewModel: MainViewModel) {
+    Column(
+        modifier = Modifier.fillMaxSize(),
+        verticalArrangement = Arrangement.Center,
+        horizontalAlignment = Alignment.CenterHorizontally
+    ) {
+        Text(text = "Hello world!")
+    }
+}
